[ feat ] 나의 이야기집 조회하기 구현

+++ app/src/main/java/com/example/simya/testData/TestDataBorder.kt
@@ -2,7 +2,7 @@ package com.example.simya.testData
 
 
 data class TestDataBorder(
-    val roomId: Int,
+    val roomId: Long,
     var todayMenu: String,
     var mainMenu: String,
     var title: String,
